Fix minification bug from styled-components
There is a bug in babel-plugin-styled-components' minification
feature. It accidentally mutates some selectors by removing
significant space.

There is selector "& :not(:last-child)" in FilterInput.baseStyles.tsx.
After minification it will turn into "&:not(:last-child)" which is
not the same selector any more.

Fixed this by disaling minification for now. I think it will take
some time until that original bug will be fixed and we can't wait
until that.