Fix popover positioning on multiselect
Popover positioning was not updated properly when using the popover
in a multiselect which is used in a container with dynamic dimensions.
For when using the multiselect in a modal, as the user selects items
from the multiselect the modal would resize and the popover would get
"detached" from the multiselect input.